scripts/gui-utility/pyproject.toml
[project]
name = "gui_utility"
version = "0.1.0"
description = "Utility scripts for gui-related things."
authors = [
    {name = "Jacob McIntosh",email = "nacitar.sevaht@gmail.com"}
]
readme = "README.md"
license = "Unlicense"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "pygobject (>=3.52.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "gui_utility", from = "src"}]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
mypy = "^1.16.1"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
skip-magic-trailing-comma = true

[tool.isort]
line_length = 79
profile = "black"

[tool.mypy]
strict = true

[tool.poetry.scripts]
app = "gui_utility.__main__:main"
checks = "gui_utility.devtools.checks:main"
